ia-image-main/requirements.txt
flask
Pillow
torch>=2.0
torchvision
numpy
matplotlib
tqdm
transformers 
diffusers
accelerate
git+https://github.com/openai/CLIP.git

